+++ kotlin-compiler-test-extensions/main/kotlin/com/javiersc/kotlin/compiler/test/runners/BoxTest.kt
@@ -15,7 +15,6 @@ import org.jetbrains.kotlin.test.builders.fir2IrStep
 import org.jetbrains.kotlin.test.builders.irHandlersStep
 import org.jetbrains.kotlin.test.builders.jvmArtifactsHandlersStep
 import org.jetbrains.kotlin.test.directives.CodegenTestDirectives.DUMP_IR
-import org.jetbrains.kotlin.test.directives.LanguageSettingsDirectives
 import org.jetbrains.kotlin.test.model.FrontendKind
 import org.jetbrains.kotlin.test.model.TestModule
 import org.jetbrains.kotlin.test.runners.RunnerWithTargetBackendForTestGeneratorMarker
@@ -35,9 +34,8 @@ public abstract class BoxTest : BaseTestRunner(), RunnerWithTargetBackendForTest
     public open fun commonServicesConfiguration(): ((FrontendKind<*>) -> Unit)? = null
 
     override fun TestConfigurationBuilder.configuration() {
-        defaultDirectives {
+        defaultDirectives { //
             +DUMP_IR
-            LanguageSettingsDirectives.LANGUAGE with "+ContextReceivers"
         }
 
         commonPluginConfiguration(